test(cudf): move infer_format regression tests to test_to_datetime.py
Address review feedback: relocate the parametrized infer_format unit test
and the to_datetime integration regression test from test_astype.py to
test_to_datetime.py, where datetime-related tests belong. No behavior
change; the module-level infer_format import moves with them.
